requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
langchain-google-genai==1.0.10
langchain-core==0.1.52
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 